move test student mode button to the students tab
