+++ src/main/resources/frontend/src/views/AdminView/AdminView.vue
@@ -13,22 +13,10 @@ import Students from "@/views/AdminView/Students.vue";
 
 useAdminStore().updateUsers();
 
-const activateTestStudentMode = async () => {
-  await testStudentModeGet()
-  window.location.href = '/';
-}
 </script>
 
 <template>
   <div class="container">
-    <div class="test-student-mode-container">
-      <p>Click on the button below to become the test student</p>
-      <p>- you will need to log out and back in again to return to admin
-        mode</p>
-      <p>- you will not be able to enter student mode unless the Test Student has a submission for the GitHub Repository
-        assignment on Canvas</p>
-      <button @click="activateTestStudentMode">Test Student Mode</button>
-    </div>
     <Tabs>
       <Tab title="Submissions">
         <SubmissionsView/>
@@ -50,14 +38,5 @@ const activateTestStudentMode = async () => {
 <style scoped>
 .container {
   min-height: 50vh;
-}
-
-.test-student-mode-container {
-  margin-bottom: 20px;
-  padding: 10px;
-  border: 1px solid #ccc;
-  border-radius: 5px;
-  background-color: #f2f2f2;
-  cursor: pointer;
 }
 </style>
